Move FileUtil and add #loadFileOrResource
