+++ even-more-fish-plugin/src/main/java/com/oheers/fish/addons/AddonManager.java
@@ -4,7 +4,7 @@ package com.oheers.fish.addons;
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.addons.impl.Head64ItemAddon;
 import com.oheers.fish.api.addons.Addon;
-import com.oheers.fish.api.addons.FileUtil;
+import com.oheers.fish.api.FileUtil;
 import com.oheers.fish.api.addons.Futures;
 import com.oheers.fish.api.addons.ItemAddon;
 import com.oheers.fish.api.addons.exceptions.JavaVersionException;
